receipt listener: fix goroutine leak
The makeUnboundedBuffered goroutine only exits once the consumer has
consumed all messages in the output channel. So instead of waiting on
the done channel in WaitForMetaTxn (which will potentially leave unread
messages), we instead rely on the open/close state of the subscription
channel to see if we are finished or not.